+++ arc-script/core/src/lib.rs
@@ -28,6 +28,7 @@ mod ssa;
 mod symbols;
 mod typer;
 mod lexer;
+mod printer;
 
 // #[cfg(feature = "shaper")]
 // mod shaper;
