+++ arc-script/core/src/printer.rs
@@ -1,0 +1,32 @@
+use crate::info::Info;
+
+pub struct Printer<'a> {
+    pub info: &'a Info<'a>,
+    pub tabs: u32,
+    pub verbose: bool,
+}
+
+const TAB: &str = "    ";
+
+impl Printer<'_> {
+    pub fn indent(&self) -> String {
+        format!("\n{}", (0..self.tabs).map(|_| TAB).collect::<String>())
+    }
+
+    pub fn tab(&self) -> Printer {
+        Printer {
+            info: self.info,
+            tabs: self.tabs + 1,
+            verbose: self.verbose,
+        }
+    }
+
+    pub fn untab(&self) -> Printer {
+        Printer {
+            info: self.info,
+            tabs: self.tabs - 1,
+            verbose: self.verbose,
+        }
+    }
+}
+
