+++ arc-script/core/src/utils.rs
@@ -1,6 +1,4 @@
-use crate::info::Info;
 use crate::prelude::*;
-use crate::symbols::SymbolTable;
 
 impl SyntaxTree {
     // Post-order traversal (Leaves first)
@@ -107,36 +105,6 @@ for_each_expr! {
     name: for_each_expr_postorder,
     pre: false,
     post: true
-}
-
-pub struct Printer<'a> {
-    pub info: &'a Info<'a>,
-    pub tabs: u32,
-    pub verbose: bool,
-}
-
-const TAB: &str = "    ";
-
-impl Printer<'_> {
-    pub fn indent(&self) -> String {
-        format!("\n{}", (0..self.tabs).map(|_| TAB).collect::<String>())
-    }
-
-    pub fn tab(&self) -> Printer {
-        Printer {
-            info: self.info,
-            tabs: self.tabs + 1,
-            verbose: self.verbose,
-        }
-    }
-
-    pub fn untab(&self) -> Printer {
-        Printer {
-            info: self.info,
-            tabs: self.tabs - 1,
-            verbose: self.verbose,
-        }
-    }
 }
 
 pub fn merge<T>(mut a: Vec<T>, mut b: Vec<T>) -> Vec<T> {
